Traits was moved to the `Concerns` directory
