+++ src/Concerns/Database.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DragonCode\LaravelActions\Traits;
+namespace DragonCode\LaravelActions\Concerns;
 
 /** @mixin \Illuminate\Database\Console\Migrations\BaseCommand */
 trait Database
