+++ src/Console/Make.php
@@ -2,10 +2,10 @@
 
 namespace DragonCode\LaravelActions\Console;
 
+use DragonCode\LaravelActions\Concerns\Argumentable;
+use DragonCode\LaravelActions\Concerns\Database;
+use DragonCode\LaravelActions\Concerns\Infoable;
 use DragonCode\LaravelActions\Constants\Names;
-use DragonCode\LaravelActions\Traits\Argumentable;
-use DragonCode\LaravelActions\Traits\Database;
-use DragonCode\LaravelActions\Traits\Infoable;
 use Illuminate\Database\Console\Migrations\BaseCommand;
 use Illuminate\Database\Migrations\MigrationCreator;
 use Illuminate\Support\Composer;
